src/test/java/seedu/task/model/task/DateTest.java
package seedu.task.model.task;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

import seedu.task.model.task.Date;

public class DateTest {

    @Test
    public void isValidDate() {
        // invalid phone numbers
        assertFalse(Date.isValidDate("")); // empty string
        assertFalse(Date.isValidDate(" ")); // spaces only
        assertFalse(Date.isValidDate("phone")); // non-numeric
        assertFalse(Date.isValidDate("9011p041")); // alphabets within digits
        assertFalse(Date.isValidDate("9312 1534")); // spaces within digits

        // valid phone numbers
        assertTrue(Date.isValidDate("93121534"));
        assertTrue(Date.isValidDate("4")); // short phone numbers
        assertTrue(Date.isValidDate("124293842033123")); // long phone numbers
    }
}
